Update README.md
This is a simple ad blocker extension for web browsers. It utilizes the "declarativeNetRequest" API to block requests to specified ad-serving domains. The manifest file defines the extension's basic properties such as name and version, along with required permissions. The "rules.json" file contains a set of rules specifying which URLs to block, such as those from doubleclick.net and googleadservices.com. When installed, the extension will automatically block requests to these domains, helping to reduce unwanted advertisements while browsing the web.